src/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.0.3
xlsxwriter==3.1.9
pydantic==2.4.2
PyJWT==2.6.0
bcrypt==4.0.1
gunicorn==21.2.0
Werkzeug==2.3.7
